+++ src/main.rs
@@ -2,7 +2,7 @@ mod card_generators;
 mod components;
 mod dynamic_content;
 
-use card_generators::placeholder_card_generator;
+use card_generators::generate_commit_card;
 use dynamic_content::{ApiRefresh, Book, Commit, Movie, Song};
 use stylist::yew::{styled_component, Global};
 use yew::prelude::*;
@@ -41,7 +41,7 @@ pub fn App() -> Html {
                         background-color: rgb(237, 244, 255);
                     }
                 "#)} />
-            { placeholder_card_generator() }
+            { generate_commit_card() }
         </div>
     }
 }
